- Created inputFiles class which processes input files and opens reading streams. - Changed project format to Gradle (from Maven).
