+++ in1.txt
@@ -1,0 +1,4 @@
+1
+4
+6
+12
