+++ in2.txt
@@ -1,0 +1,4 @@
+22
+46
+68
+80
