+++ in3.txt
@@ -1,0 +1,4 @@
+44
+62
+88
+101
